requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
customtkinter==5.2.2
darkdetect==0.8.0
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
numpy==2.2.5
packaging==25.0
pillow==11.2.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
urllib3==2.4.0
